Correção noninteractive install

+++ install.sh
@@ -69,25 +69,26 @@ init_install() {
 
   print_message 'Ferramenta em script Bash Completa para Bug bounty ou Pentest ! Vai poupar seu Tempo na hora de configurar sua máquina para trabalhar.'
   printf "\n${CBold}${CFGWhite}=====================================================>${CReset}\n\n"
-  print_message 'Deseja Atualizar seu Linux? o tempo pode variar de acordo com sua máquina.'
-  PS3="Por favor selecione uma opção : "
-  read -p "$PS3"
-  select opt in yes no; do
-    case $opt in
-      yes)
-        printf '\natualizando..\n'
-        apt -y full-upgrade
-        sudo $SUDO_OPT pip3 install --upgrade pip
-        sudo $SUDO_OPT pip3 install --upgrade osrframework
-        apt -y autoremove
-        break
-        ;;
-      no) print_message 'continuando com a instalação...'
-        break
-        ;;
-      *) printf '\nOpção inválida\n'
-    esac
-  done
+  if [[ -t 0 ]]; then
+    print_message 'Deseja Atualizar seu Linux? o tempo pode variar de acordo com sua máquina.'
+    PS3="Por favor selecione uma opção: "
+    select opt in yes no; do
+      case $opt in
+        yes)
+          printf '\natualizando..\n'
+          apt -y full-upgrade
+          sudo $SUDO_OPT pip3 install --upgrade pip
+          sudo $SUDO_OPT pip3 install --upgrade osrframework
+          apt -y autoremove
+          break
+          ;;
+        no) print_message 'continuando com a instalação...'
+          break
+          ;;
+        *) printf '\nOpção inválida\n';;
+      esac
+    done
+  fi
 }
 
 progressbar() {
@@ -190,6 +191,7 @@ if [[ $# == 0 ]]; then
   selection="${!tools[*]}"
 fi
 
+export DEBIAN_FRONTEND=noninteractive
 init_install
 for tool in ${selection,,}; do
   tool_list=${!tools[*]}
